# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Food Buddy Kitchen Intelligence API
  description: >-
    AI-powered kitchen intelligence — scan ingredients, manage a pantry, get
    pantry-based recipe recommendations, ingredient substitutions, meal
    planning, and the Buddy agentic assistant.


    Note: the real-time Buddy assistant is also exposed over a WebSocket at
    `GET /ws/buddy` (not expressible in OpenAPI). Clients send JSON messages
    like `{"trigger": "user_action"}` and receive `BuddyEvent` payloads.


    A legacy-prefix middleware maps unprefixed paths (e.g. `/pantry`, `/recipes`,
    `/scan`, `/substitute`, `/planner`, `/buddy`) to their `/api/...`
    equivalents.
  version: 1.0.0
  license:
    name: Proprietary
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://rasoi-backend-ml4i.onrender.com
    description: Render production
...
paths:
...
  /api/scan:
    post:
      tags: [scan]
      summary: Scan an ingredient photo or grocery receipt
      operationId: scanImage
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image]
              properties:
                image:
                  type: string
                  format: binary
                  description: JPEG/PNG/WebP image, max 5 MB
                scanType:
                  type: string
                  default: ingredient
                  enum: [ingredient, receipt]
                append:
                  type: boolean
                  default: false
                  description: false = replace pantry, true = add to existing
                userId:
                  type: string
                  default: guest
                knownExpirationDate:
                  type: string
                  format: date
                  nullable: true
      responses:
        '200':
          description: Extracted and stored ingredients
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ScanResponse' }
        '400':
          description: Invalid file type/size or bad form field
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '502':
          description: Vision analysis failed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '503':
          description: AI client not configured (missing ANTHROPIC_API_KEY)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/pantry:
    get:
      tags: [pantry]
      summary: List all pantry items
      operationId: getPantry
      responses:
        '200':
          description: All items sorted by expiry (soonest first)
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/PantryItem' }
    post:
      tags: [pantry]
      summary: Add a pantry item
      operationId: addPantryItem
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PantryItemCreateRequest' }
      responses:
        '200':
          description: Item created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  item: { $ref: '#/components/schemas/PantryItem' }
...
  /api/recipes:
    get:
      tags: [recipes]
      summary: Meal recommendations from the current pantry
      operationId: getRecipes
      parameters:
        - name: prioritize_expiring
          in: query
          schema: { type: boolean, default: true }
        - name: max_recipes
          in: query
          schema: { type: integer, minimum: 1, maximum: 20, default: 5 }
        - name: cuisine
          in: query
          schema: { type: string }
          description: e.g. any, Indian, South Indian, North Indian, Pan Indian, Italian, Mexican
        - name: meal_type
          in: query
          schema: { type: string }
        - name: diet
          in: query
          schema: { type: string }
        - name: max_ready_time
          in: query
          schema: { type: integer, minimum: 1, maximum: 240 }
      responses:
        '200':
          description: Recipe recommendations
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RecipesResponse' }
        '422':
          description: Invalid query parameter
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /api/substitute:
    post:
      tags: [substitutions]
      summary: Suggest pantry-available substitutes for a missing ingredient
      operationId: substitute
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SubstitutionRequest' }
      responses:
        '200':
          description: Substitution suggestions
          content:
            application/json:
              schema:
                type: object
                properties:
                  substitutions:
                    type: array
                    items: { $ref: '#/components/schemas/Substitution' }
        '400':
          description: Empty/invalid input
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '502':
          description: Substitution lookup failed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '503':
          description: AI client not configured
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
components:
...
  schemas:
    Error:
      type: object
      properties:
        detail: { type: string }
        message: { type: string }
    PantryItem:
      type: object
      properties:
        id: { type: string }
        name: { type: string, example: tomato }
        quantity: { type: number, example: 4 }
        unit: { type: string, example: pcs }
        acquisitionDate: { type: string, format: date }
        expirationDate: { type: string, format: date }
        createdAt: { type: string }
        updatedAt: { type: string }
        isExpiring: { type: boolean, description: expiring within 2 days }
        isExpired: { type: boolean }
    PantryItemCreateRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        quantity: { type: number, default: 1.0 }
        unit: { type: string, default: pcs }
        acquisitionDate: { type: string, format: date, nullable: true }
        expirationDate: { type: string, format: date, nullable: true }
...
    ScanResponse:
      type: object
      properties:
        success: { type: boolean }
        ingredients:
          type: array
          items: { $ref: '#/components/schemas/PantryItem' }
        message: { type: string }
        imageUrl: { type: string, nullable: true }
    RecipeIngredient:
      type: object
      properties:
        name: { type: string }
        quantity: { oneOf: [ { type: string }, { type: number } ] }
        unit: { type: string }
        available: { type: boolean }
        is_optional: { type: boolean }
        sort_order: { type: integer }
    RecipeStep:
      type: object
      properties:
        step_number: { type: integer }
        instruction: { type: string }
        duration_min: { type: integer, nullable: true }
        tip: { type: string, nullable: true }
    Recipe:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        name: { type: string }
        cuisine: { type: string, nullable: true }
        meal_type: { type: string, nullable: true }
        diet: { type: string, nullable: true }
        difficulty: { type: string, enum: [Easy, Medium, Hard], nullable: true }
        prepTimeMinutes: { type: integer }
        ready_in_min: { type: integer, nullable: true }
        servings: { type: integer, nullable: true }
        image_url: { type: string, nullable: true }
        description: { type: string, nullable: true }
        calories_kcal: { type: integer, nullable: true }
        protein_g: { type: number, nullable: true }
        carbs_g: { type: number, nullable: true }
        fat_g: { type: number, nullable: true }
        fiber_g: { type: number, nullable: true }
        matchPercentage: { type: number }
        usesExpiringItems: { type: boolean }
        missingIngredients:
          type: array
          items: { type: string }
        ingredients:
          type: array
          items: { $ref: '#/components/schemas/RecipeIngredient' }
        steps:
          type: array
          items: { type: string }
        cooking_steps:
          type: array
          items: { $ref: '#/components/schemas/RecipeStep' }
        source: { type: string, enum: [supabase, spoonacular, openai] }
    RecipesResponse:
      type: object
      properties:
        success: { type: boolean }
        provider: { type: string, example: openai }
        message: { type: string }
        recipes:
          type: array
          items: { $ref: '#/components/schemas/Recipe' }
...
    SubstitutionRequest:
      type: object
      required: [recipeId, missingIngredient, recipeContext]
      properties:
        recipeId: { type: string }
        missingIngredient: { type: string }
        recipeContext: { type: string, description: recipe name for prompt context }
    Substitution:
      type: object
      properties:
        ingredient: { type: string }
        ratio: { type: string, example: '1:1' }
        notes: { type: string }
        available: { type: boolean }
...